moved the fetch to admin page
