+++ client/src/pages/Admin.js
@@ -1,7 +1,20 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React from "react";
+import { useState,useEffect } from "react";
 
 function AdminPage() {
+  const [pages, setPages] = useState([]);
+
+  async function fetchPagesAPI() {
+    const page = await fetch("http://localhost:3000/api/home");
+    const response = await page.json();
+    setPages(response);
+	
+  }
+ 
+  useEffect(() => {
+     fetchPagesAPI();
+  }, []);
 	return (
 		<div>
             <h1>Page 1</h1>
